python/format/pyproject.toml
[project]
name = "format"
description = "MLCommons datasets format."
version = "0.0.1"
authors = [
  { name = "Joaquin Vanschoren" },
  { name = "Jos van der Velde" },
  { name = "Omar Benjelloun" },
  { name = "Peter Mattson" },
  { name = "Pieter Gijsbers" },
  { name = "Pierre Marcenac" },
  { name = "Pierre Ruyssen" },
  { name = "Prabhant Singh" },
]
dependencies = ["absl-py", "etils[epath]", "networkx", "rdflib"]
readme = "README.md"

[project.optional-dependencies]
dev = ["pytest"]

[tool.black]
line-length = 80
